oatmeal/backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-mssql-backend>=2.8.1
pyodbc>=4.0.34
python-dotenv>=1.0.0 